implementing count and linting

+++ internal/structs/api_structs.go
@@ -21,13 +21,13 @@ type Query struct {
 }
 
 type PayloadsData struct {
-	Count   int               `json:"count"`
+	Count   int64             `json:"count"`
 	Elapsed float64           `json:"elapsed"`
 	Data    []models.Payloads `json:"data"`
 }
 
 type ErrorResponse struct {
-	Title	string		`json:"title"`
-	Message	string		`json:"message"`
-	Status	int			`json:"status"`
+	Title   string `json:"title"`
+	Message string `json:"message"`
+	Status  int    `json:"status"`
 }
